+++ packages/pn-pa-webapp/src/__mocks__/NotificationDetail.mock.ts
@@ -1062,6 +1062,7 @@ export const recipients: Array<NotificationDetailRecipient> = [
 export const notificationDTOMultiRecipient: NotificationDetail = {
   abstract: 'Abstract della notifica',
   subject: 'notifica analogica con cucumber',
+  paProtocolNumber: '302125256777757822',
   recipients,
   filedAt: '2023-08-25T09:33:58.709695008Z',
   documents: [