+++ packages/pn-personafisica-webapp/src/navigation/routes.const.ts
@@ -37,7 +37,6 @@ export const DIGITAL_DOMICILE = `${RECAPITI}/domicilio-digitale`;
 export const DIGITAL_DOMICILE_ACTIVATION = `${DIGITAL_DOMICILE}/attivazione`;
 export const DIGITAL_DOMICILE_MANAGEMENT = `${DIGITAL_DOMICILE}/gestione`;
 export const LOGOUT = '/auth/logout';
-export const LOGOUT_OI = '/auth/logout-oi';
 export const TPP_LANDING = '/nuova-notifica-send';
 export const ONBOARDING = '/onboarding';
 export const ONBOARDING_DIGITAL_DOMICILE = 'domicilio-digitale';