+++ packages/pn-personagiuridica-webapp/src/pages/NotificationDetail.page.tsx
@@ -508,7 +508,7 @@ const NotificationDetail = () => {
                   filedAt={notification.filedAt}
                   iun={notification.iun}
                   abstract={notification.abstract}
-                  senderLogoUrl={SELFCARE_CDN_URL}
+                  selfcareCdnUrl={SELFCARE_CDN_URL}
                 />
                 {banner}
                 {pecUnreachableAlert}